renamed files for clarity, refactored some code
